feat: セキュリティコンフィグをMastodon互換APIと分離
